requirements.txt
streamlit
chromadb
google-generativeai
python-dotenv
astunparse
tiktoken
tree-sitter
psutil
sentence_transformers
scikit-learn
